Forgot Password button opens password reset form (Web Browser only)
Added a simple function to the forgot password button on the login screen. For now it opens the built in django password recovery form and allows the user to change their password. It only works on web browsers at the moment since the django server is run on the local machine and expo apps running on the phone can't connect to it.